# batches: Faster listing of changesets by type